+++ packages/fiori-app-sub-generator/src/types/state.ts
@@ -11,6 +11,7 @@ import type {
 } from '@sap-ux/odata-service-inquirer';
 import type { ApiHubType, SapSystemSourceType } from '../types/constants';
 import type { Script } from './common';
+import type { AppGenInfo } from '@sap-ux/fiori-generator-shared';
 
 export interface Project {
     targetFolder: string;
@@ -125,6 +126,10 @@ export interface State {
      * The template selected by the user, maybe FE or FF
      */
     floorplan: Floorplan;
+    /**
+     * General information about the application - used for README and appGenInfo.json
+     */
+    appGenInfo?: Partial<AppGenInfo>;
 }
 
 /** Only used by headless */