+++ packages/adp-flp-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,63 @@
 # @sap-ux/adp-flp-config-sub-generator
 
+## 0.1.9
+
+### Patch Changes
+
+-   @sap-ux/adp-tooling@0.14.9
+-   @sap-ux/flp-config-inquirer@0.3.9
+
+## 0.1.8
+
+### Patch Changes
+
+-   Updated dependencies [ac55cca]
+    -   @sap-ux/fiori-generator-shared@0.12.2
+    -   @sap-ux/flp-config-inquirer@0.3.8
+    -   @sap-ux/inquirer-common@0.7.3
+    -   @sap-ux/adp-tooling@0.14.8
+
+## 0.1.7
+
+### Patch Changes
+
+-   Updated dependencies [b3fe5b8]
+    -   @sap-ux/inquirer-common@0.7.2
+    -   @sap-ux/adp-tooling@0.14.7
+    -   @sap-ux/flp-config-inquirer@0.3.7
+
+## 0.1.6
+
+### Patch Changes
+
+-   @sap-ux/adp-tooling@0.14.6
+-   @sap-ux/flp-config-inquirer@0.3.6
+
+## 0.1.5
+
+### Patch Changes
+
+-   Updated dependencies [2224d63]
+    -   @sap-ux/axios-extension@1.21.1
+    -   @sap-ux/adp-tooling@0.14.5
+    -   @sap-ux/system-access@0.6.1
+    -   @sap-ux/flp-config-inquirer@0.3.5
+
+## 0.1.4
+
+### Patch Changes
+
+-   Updated dependencies [8e44d02]
+    -   @sap-ux/adp-tooling@0.14.4
+    -   @sap-ux/flp-config-inquirer@0.3.4
+
+## 0.1.3
+
+### Patch Changes
+
+-   @sap-ux/adp-tooling@0.14.3
+-   @sap-ux/flp-config-inquirer@0.3.3
+
 ## 0.1.2
 
 ### Patch Changes