+++ packages/app-config-writer/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/app-config-writer
 
+## 0.6.4
+
+### Patch Changes
+
+-   f8deed9: Remove '@sap-ux/cards-editor-middleware' dependency and 'sap-cards-generator' custom middleware for application having card generation enabled already.
+
+## 0.6.3
+
+### Patch Changes
+
+-   Updated dependencies [2224d63]
+    -   @sap-ux/axios-extension@1.21.1
+
 ## 0.6.2
 
 ### Patch Changes